python/requirements.txt
playwright==1.43.0
python-dotenv==1.0.1
browserbase==1.2.0
